# openapi/pnpu-portal.openapi.yml
openapi: 3.1.0
info:
  title: PNPU Portal API
  version: 0.1.0
  description: API técnica inicial del Portal PNPU.
servers:
  - url: http://127.0.0.1:4307
    description: Desarrollo local
paths:
  /v1/publications:
    get:
      operationId: listPublications
      summary: Lista publicaciones públicas del catálogo PNPU.
      tags:
        - Catalog
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PublicationSearch"
        - $ref: "#/components/parameters/PublicationPublisherId"
        - $ref: "#/components/parameters/PublicationContributorId"
        - $ref: "#/components/parameters/PublicationCollectionId"
        - $ref: "#/components/parameters/PublicationLanguage"
        - $ref: "#/components/parameters/PublicationSubject"
        - $ref: "#/components/parameters/PublicationSort"
      responses:
        "200":
          description: Listado paginado de publicaciones.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationListResponse"
        "422":
          description: Parámetros de paginación inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/publications/{id}:
    get:
      operationId: getPublication
      summary: Devuelve la ficha pública de una publicación por UUID PNPU.
      tags:
        - Catalog
      parameters:
        - name: id
          in: path
          required: true
          description: UUID v7 público de la publicación.
          schema:
            $ref: "#/components/schemas/PnpuUuid"
      responses:
        "200":
          description: Ficha pública de la publicación.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationDetailResponse"
        "404":
          description: Publicación no encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Identificador inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    PublicationImportToken:
      type: apiKey
      in: header
      name: X-PNPU-Admin-Token
    PublicationImportOidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: Número de página, empezando en 1.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: Cantidad de elementos por página.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    PublicationSearch:
      name: q
      in: query
      required: false
      description: Texto de búsqueda sobre título, subtítulo, resumen, editorial, identificadores, materias y palabras clave.
      schema:
        type: string
        maxLength: 120
    PublicationPublisherId:
      name: publisherId
      in: query
      required: false
      description: UUID PNPU de la editorial.
      schema:
        $ref: "#/components/schemas/PnpuUuid"
    PublicationContributorId:
      name: contributorId
      in: query
      required: false
      description: UUID PNPU del autor o contribuyente.
      schema:
        $ref: "#/components/schemas/PnpuUuid"
    PublicationCollectionId:
      name: collectionId
      in: query
      required: false
      description: UUID PNPU de la colección editorial.
      schema:
        $ref: "#/components/schemas/PnpuUuid"
    PublicationLanguage:
      name: language
      in: query
      required: false
      description: Idioma ISO 639-1 de la publicación.
      schema:
        type: string
        pattern: ^[a-z]{2}$
    PublicationSubject:
      name: subject
      in: query
      required: false
      description: Texto o identificador de materia.
      schema:
        type: string
        maxLength: 120
    PublicationSort:
      name: sort
      in: query
      required: false
      description: Criterio de ordenamiento del listado de publicaciones.
      schema:
        type: string
        default: publicationDateDesc
        enum:
          - publicationDateDesc
          - publicationDateAsc
          - titleAsc
          - titleDesc
          - publisherAsc
  schemas:
    PnpuUuid:
      type: string
      pattern: ^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-7[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}$
    ApiMeta:
      type: object
      additionalProperties: false
      required:
        - apiVersion
      properties:
        apiVersion:
          type: string
          const: v1
    ApiLinks:
      type: object
      additionalProperties: false
      required:
        - self
      properties:
        self:
          type: string
          format: uri
        next:
          type: string
          format: uri
        prev:
          type: string
          format: uri
    Pagination:
      type: object
      additionalProperties: false
      required:
        - page
        - pageSize
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    PublisherSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - officialName
        - country
      properties:
        id:
          $ref: "#/components/schemas/PnpuUuid"
        officialName:
          type: string
        acronym:
          type: string
        country:
          type: string
          pattern: ^[A-Z]{2}$
...
    ContributorSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - roles
      properties:
        id:
          $ref: "#/components/schemas/PnpuUuid"
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        orcid:
          type: string
          format: uri
...
    Identifier:
      type: object
      additionalProperties: false
      required:
        - type
        - value
      properties:
        type:
          type: string
          enum:
            - isbn
            - eisbn
            - doi
            - uri
            - handle
            - pnpuUuid
        value:
          type: string
    SubjectSummary:
      type: object
      additionalProperties: false
      required:
        - identifier
        - preferredLabel
      properties:
        identifier:
          type: string
        preferredLabel:
          type: string
        uri:
          type: string
          format: uri
...
    Resource:
      type: object
      additionalProperties: false
      required:
        - type
        - url
        - format
      properties:
        type:
          type: string
        url:
          type: string
          format: uri
        format:
          type: string
        fileSize:
          type: integer
          minimum: 0
        checksum:
          type: string
        language:
          type: string
          pattern: ^[a-z]{2}$
        license:
          type: string
    PublicationSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - publicationDate
        - language
        - type
        - subjects
        - publisher
      properties:
        id:
          $ref: "#/components/schemas/PnpuUuid"
        title:
          type: string
        subtitle:
          type: string
        publicationDate:
          type: string
          format: date
        language:
          type: string
          pattern: ^[a-z]{2}$
        type:
          type: string
        license:
          type: string
        primaryIdentifier:
          $ref: "#/components/schemas/Identifier"
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/SubjectSummary"
        keywords:
          type: array
          items:
            type: string
        publisher:
          $ref: "#/components/schemas/PublisherSummary"
    PublicationDetail:
      allOf:
        - $ref: "#/components/schemas/PublicationSummary"
        - type: object
          additionalProperties: false
          required:
            - format
            - contributors
            - identifiers
            - subjects
            - resources
          properties:
            subtitle:
              type: string
            abstract:
              type: string
            format:
              type: string
            license:
              type: string
            contributors:
              type: array
              items:
                $ref: "#/components/schemas/ContributorSummary"
            identifiers:
              type: array
              items:
                $ref: "#/components/schemas/Identifier"
            subjects:
              type: array
              items:
                $ref: "#/components/schemas/SubjectSummary"
            resources:
              type: array
              items:
                $ref: "#/components/schemas/Resource"
            keywords:
              type: array
              items:
                type: string
    PublicationListResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - pagination
        - links
        - meta
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PublicationSummary"
        pagination:
          $ref: "#/components/schemas/Pagination"
        links:
          $ref: "#/components/schemas/ApiLinks"
        meta:
          $ref: "#/components/schemas/ApiMeta"
    PublicationDetailResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - links
        - meta
      properties:
        data:
          $ref: "#/components/schemas/PublicationDetail"
        links:
          type: object
          additionalProperties: false
          required:
            - self
          properties:
            self:
              type: string
              format: uri
        meta:
          $ref: "#/components/schemas/ApiMeta"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - correlationId
      properties:
        code:
          type: string
          enum:
            - PNPU-404
            - PNPU-403
            - PNPU-409
            - PNPU-422
            - PNPU-429
            - PNPU-503
        message:
          type: string
        correlationId:
          type: string
...